allow kafka poll tuning
Previously, we used the kafka poll defaults for the number of records
to pull (500), and how long to wait to process said records (5 min).

This seems to work fine for archive processing but does not work for
deletes. Deleting 500 inventory IDs can take longer than 5 minutes in
some cases. When this happens, our kafka client times out and puts the
delete messages back on the queue for reprocessing. This can result in
a pile-up of messages since we are unsuccessfully reprocessing the
same messages over and over.

This commit exposes two options: `KAFKA_MAX_POLL_INTERVAL_MS` (default
5 min) and `KAFKA_MAX_POLL_RECORDS` (default 500). We also now log the
  values for these at startup.