do not save multiple profiles with the same date
If a reporter updates an inventory record, it will fire an "updated
record" message on the egress topic. However, not all of these updates
actually update the system profile. This results in duplicate system
profiles in the historical DB.

This change to the archiver checks to ensure a record does not already
exist before saving it to the DB.

We could do this more elegantly via a DB uniqueness constraint;
however this way does not require a data migration to clean old
records. We may eventually want to add a constraint once the data is
unique.

Note that this does not do the select and insert in the same
transaction.  I don't anticipate issues arising from this since the
"fail" case (which should be rare) is that we insert two records which
is the current behavior.  Adding a constraint would also let us avoid
having to add a transaction with a select and insert.